refactor: 폴더구조 변경(기능별 분리)
