--- src/pages/Repair.jsx
@@ -1,5 +1,0 @@
-import RepairTemplate from '../templates/RepairTemplate';
-
-export default function Repair() {
-  return <RepairTemplate />;
-}
