+++ src/templates/repair/request-repair/RequestRepairTemplate.jsx
@@ -4,11 +4,11 @@ import styled, { css } from 'styled-components';
 import { useFunnel } from '@use-funnel/react-router-dom';
 import { useNavigate } from 'react-router-dom';
 
-import TopBar from '../components/common/TopBar';
-import ModeItem from '../components/common/ModeItem';
-import Button from '../components/common/Button';
-import { Row, Column, Spacer } from '../styles/flex';
-import { color, typo } from '../styles/tokens';
+import TopBar from '../../../components/common/TopBar';
+import ModeItem from '../../../components/common/ModeItem';
+import Button from '../../../components/common/Button';
+import { Row, Column, Spacer } from '../../../styles/flex';
+import { color, typo } from '../../../styles/tokens';
 
 /* =========================================================
  * 공통 상수/유틸
